Short-circuit test execution if the chain has stopped finalizing
instead of blindly pushing forward which causes CI jobs to take hours
and hours of execution time and clog the pipeline.

Add a global beforeEach() hook which executes *before* any of the
hooks defined for each test scenario. This should help us detect
situations where a newly added test messes up the environment and
makes the chain to not finalize blocks anymore!

Add a helper function to exercise the assertions.