Rename internal variables
it looks like that getValidatorStatus() is always being called correctly
with the address of an expected validator regardless of whether or not
we are using proxy accounts.

This change just makes all of the invocations consistent with one
another in case we'd like to somehow assert/grep on these calls in the
future!