Do not force new eras during tests
when calling waitEras() during the integration tests we shouldn't be
forcing new eras by default because this is not what the standard chain
behvior is.

For the situation where we know what we're doing and really want to
force a new era there is the function `forceNewEra()` which the scenario
should be calling directly!