Conditionally execute validatorCycle.test.ts scenario
it takes over 10 minutes with a max timeout of around 30 mins
and is generally very slow while all of the commands involved in this
scenario are also exercised independently as part of the rest of the
test suite.

This test will be executed conditionally only for commits on `dev` branch
but not on every single PR. Will also be executed for PRs against `testnet`
(e.g. before a release).