Partial revert of 'Remove self-hosted runner as soon as we're done with it'
this partially reverts the previous commit because it simply does not
work!

Some GHA actions have a post step which executes *after* we start the
deletion process in Linode. For first time PRs compressing and uploading
Rust cache usually takes a few minutes in which time the runner VM is
already gone. This causes a failure which in turn causes dependent jobs
to not be triggered.

* Linode does not support sheduling deletions in their API
* (sleep 600; .github/remove-linode-vm.sh) &  doesn't work either
  because by the time the background job should execute the runner has
  completed the current job and has self-unregistered from GitHub, most
  likely also killing any background processes
- detaching it from the terminal and running in background doesn't work
  either and it's impossible to debug. I suspect that at least the
  environment isn't passed down properly

Best effort we can do is to schedule a system shutdown to avoid extra
billing until the VM is removed when the entire workflow finishes. Even
that is UNRELIABLE and sometimes the VM stays running when not needed!