Decouple self-hosted removal jobs from CI matrix
specify each job independently instead of as a matrix in order to be able
to remove self-hosted runners as soon as they are done instead of
waiting for all possible jobs in the matrix to finish!

NOTE: this necessitates that we have lots of repeated job definitions in
the ci.yml file b/c there's no other way around it.

NOTE: some removals are still groupped together because their
corresponding jobs take roughly the same amount of time.