Place ClaimProver files in /var/tmp/creditcoin3/claim-proofs/
the call to .with_default_files() happens via
cairo_generate_proof() -> build_prover() which is called for specific
block as an argument and the directory structure is explicitly created
every time!

Therefore we can place these files in a central location which we know
exists on every Linux OS and avoid the use of relative paths!