Adjust Cairo PATH inside container & remove redundant binaries
now that we rely on PATH to be configured so that all Cairo components
can be found we adjust the PATH inside the container to include all
necessary directories. Which also means that we don't really need the
files `/bin/cpu_air_prover` and `/bin/cpu_air_verifier` anymore because
they are also available under `/cairo/stone-*/cpu_air_*` inside the
container!