Remove ENV variable from CI matrix - defaults to EPHEMERAL=true
I believe this was introduced by mistake in gluwa/creditcoin3 and then
copied over into this repository b/c it doesn't make sense to use
non-ephemeral runners here. To clarify:

Ephemeral (true) runners are unregistered from GitHub automatically after
they execute a single job which is exactly how they are used in this
scenario. It doesn't make sense to keep them around because we're not
using them anyway. In contrast this is different for the benchmarks and
runtime-upgrade workflows where a single runner is used for several jobs
in sequence.