Use self-hosted runners when building the release
b/c this is a private repository and we are going to burn through the
available CI minutes in no time if we try building a release on GitHub
hosted runners!

NOTE: we have to install the zip binary b/c it isn't installed by
default, otherwise the thedoctor0/zip-release action will fail!