Specify dependency on deploy-github-runner CI job
Reasoning:

https://github.com/gluwa/creditcoin3-next/actions/runs/11496787579/job/31999114993?pr=264
has timed out as designed at 13:01 because it wasn't able to provision a VM for some
reason. However the clean-up part was not started immediately. Instead
it was started at 14:47 after maually clicking "Cancel workflow"!

The reason that cleanup was not started is that the old order of
dependencies was:

- <some job> <- cleanup
- deploy-github-runner

where <some job> is automatically scheduled by GitHub and just waits for
a specific runner to become available (which it doesn't). When <some job>
times out then GitHub will start all other jobs which depend on it which
currently takes at least 24 hrs and we have no way of controling it:
https://github.com/actions/actions-runner-controller/discussions/3780

With this change applied the dependency graph becomes:

deploy-github-runner <- <some job> <- cleanup

When deployment fails/times out, then <some job> will also
fail and afterwards GitHub will move onto cleanup!