Pin linode-cli version to 5.52.0 b/c of bug with latest version
at the same time specify what we want to install in requirements.txt
file and configure that via Dependabot b/c it will at least inform us
when a new dependency is available and in the case of bugs, like now,
the provisioning should fail only on a single PR instead of failing for
everyone.

NOTE: use pip instead of pipx to install this b/c pipx does not support
requirements files.