ci: Workaround the fact that we cannot sync CC3next with Creditcoin3
this used to work until recently, however it broke. When starting
the container we get:

```
2024-12-10 12:18:27 Essential task `attestor-gossip` failed. Shutting down service.
Error: Service(Other("Essential task failed."))
```

For more context see:
https://gluwa.slack.com/archives/C03MQ532BGA/p1733811879170669

This commit launches a container with a stand-alone dev network so that
we can proceed inspecting all of the other bits inside the container,
like binaries, Cairo, etc.

IMPORTANT: this should be reverted once we have CC Next Devnet and we
point the synchronization to its boot node!