Add local .eslint as symlink to cli/
resolves the error:

ESLint: 8.57.1

ESLint couldn't find the plugin "@typescript-eslint/eslint-plugin".

(The package "@typescript-eslint/eslint-plugin" was not found when loaded as a Node module from the directory "/home/runner/work/creditcoin3-next/creditcoin3-next/cli".)

It's likely that the plugin isn't installed correctly. Try reinstalling by running the following:

    npm install @typescript-eslint/eslint-plugin@latest --save-dev

IMPORTANT:

Although the corrent npm package gets installed in the current directory
it appears that eslint is searching for its plugins relevant to the
directory where the config file is located, hence the error, and not
relevant to the current working directory!