Disable unattended-upgrades.service & upgrade self-hosted runner explicitly
this applies all the latest upgrades, including security patches
(which do exist at this time) and prevents the machine from accidentally
being rebooted (which is not supposed to happen but IDK how this service works).

It is also possible that applying some of the upgrades *AFTER*
action-runner starts may have been causing restart of service on which the
application depends which could have been causing a temporary disconnect with
GitHub and triggering our CI jobs to cancel. Should be aleviated by an
explicit upgrade *BEFORE* we start action-runner!

Not sure if any of that is true but at least sounds plausible. I guess
we'll have to apply those changes and monitor if any jobs get cancelled.