Add more test scenarios for getQueryResultSegments()
simulate the 2 cases where the underlying get_result_segments()
precompile either reverts directly or produces a different kind of
error.

NOTE:
In Solidity any sort of error appears to trigger a reversal, even when
applied to a 'view' or 'pure' function. There doesn't appear to be a
distinction between an regular error and an error of type revert.

Such distinction appears to only be present on the Rust side for
precompiles:
  - if the error returned by a precompile is of type PrecompileFailure::Revert
    then that triggers a proper revert in the EVM and client-side is
    able to detect it (e.g. we can assert on it appropriately in tests)
  - when the error returned by a precompile is of different type then
    the client side sees a CALL_EXCEPTION failure without any particular
    details.

Related: CSUB-1552