Remove redundant VerifierExitStatus statuses & if block in submitQueryProof()
because the precompile now causes a revert() it doesn't actually revert
non-zero (e.g. non-Success) statuses anymore. Checking for them is a
no-op (b/c of the revert that would preceed the check) and all of this
can be removed to simplify the contract code.

I've added comments for reference in case someone in the future is
wondering why we don't check for errors.