Adjust comparison in hasBondedEnough() to compare BigInt numbers
even when comparing values which are both represented in micro-units
using the expression:

amount < minValidatorBond

we get the wrong result.

Both arguments are of type BN and for example for values:
   min=100000000000000000000
amount= 50000000000000000000

amount < min results in `false`. My guess is that these values are
internally converted to strings when comparing them.

Wrap the string representation around BigInt() so that values are
compared correctly