Adjust attestor withdraw-unbonded test scenarios to work with proxy
all scenarios which require the unbonding period to be finished are now
groupped together, even the ones where the proxy account is invalid or
has no funds. The reason being is that first we check whether or not the
address has any funds eligible for withdrawal and afterwards we try to
perform the action which either succeeds or fails with a transaction
error from the chain.