Connect to the RPC api in the only scenario which uses it directly
Speculation: the first thing this test file does is initiate an RPC
connection to creditcoin3-node over a WebSocket and then it doesn't
quite use it for most of the time. When Jest comes around to execute the
last scenario we get a WS connection error b/c creditcoin3-node has
already closed the connection.

Instead of connecting to WS in beforeAll() do that immediately before
using the `api` variable. CLI() calls initiate their own internal RPC
connection and don't share this variable. The rest of the test scenarios
in this file aren't affected b/c they don't query or modify the chain
directly, just execute the command line.

This should workaround the fact that creditcoin3-node may close the WS
connection due to inactivity documented in CSUB-1708.