Refactor internall call to .wait() inside verify() test
NOTE: this affects only the negative scenario however it apspears that
the older syntax could be wrong.

NOTE 2: I am still not convinced that this is
the way to go but it appears to work although none of the other test
scenarios (e.g. substrate_transfer(), nor the removed
get_result_segments()) uses this very same syntax - instead they call
the precompile function directly inside of expect() and assert on the
error string!