Properly encode empty proof for testing
with this change I can see in the logs:

2025-07-17 15:28:55 Empty proof submitted

and in the Substrate portal the revert reason is "Invalid proof submitted".

Previously it was a parsing error:

2025-07-17 15:27:35 error parsing the proof: Error("expected value", line: 1, column: 1)
2025-07-17 15:27:35 Proof verification failed: ProcessError

My speculation is that this parsing error was not handled correctly and
was causing the transaction to be invalid, causing all of the timeouts
and issues we were seeing.

NOTE: updates all other places where we use an "empty proof" to avoid
copy & paste errors in the future in case the actual value matters.