ci: Add TODO step documenting suggestions how to fix git conflicts
This isn't 100% bullet proof b/c it depends on what exaclty caused the
git conflict in the first place and how soon did we notice! It should
work reasonably well for gluwa-bot commits which update weights and/or
type definitions since these are usually the source of conflicts we've
seen in the past!