Update actions-runner to 2.329.0
actions/checkout@v5 requires a min of of v2.327.1 to run b/c it has
switched to node24 runtime!

This is probably the root cause of this error:

```
Error: System.ArgumentOutOfRangeException: Specified argument was out of the range of valid values. (Parameter ''using: node24' is not supported, use 'docker', 'node12', 'node16' or 'node20' instead.')
   at GitHub.Runner.Worker.ActionManifestManager.ConvertRuns(IExecutionContext executionContext, TemplateContext templateContext, TemplateToken inputsToken, String fileRelativePath, MappingToken outputs)
   at GitHub.Runner.Worker.ActionManifestManager.Load(IExecutionContext executionContext, String manifestFile)
Error: Failed to load actions/checkout/v5/action.yml
```

https://github.com/gluwa/creditcoin3/actions/runs/19126519830/job/54657980703