ci: Pipe diff output to colordiff to prevent execution short-circuit
diff will produce a non-zero exit status if the inputs are different
which short-circuits the execution of this step and doesn't print the
actual diff so we can see it.

Piping the entire content through colordiff takes care of the exit code
and should work fine b/c later we explicitly check whether diff produced
a non-empty output!